backend/requirements.txt
fastapi
slowapi
uvicorn[standard]
python-dotenv
python-multipart
requests
mutagen
